+++ dep/acelite/ace/Process.inl
@@ -1,6 +1,6 @@
 // -*- C++ -*-
 //
-// $Id: Process.inl 83405 2008-10-23 06:30:06Z johnnyw $
+// $Id: Process.inl 97660 2014-03-17 09:49:33Z johnnyw $
 
 #include "ace/ACE.h"
 #include "ace/OS_NS_sys_wait.h"
@@ -48,8 +48,7 @@ ACE_Process::gethandle (void) const
 }
 
 ACE_INLINE pid_t
-ACE_Process::getpid (void)
-    const
+ACE_Process::getpid (void) const
 {
 #if defined (ACE_WIN32)
   return process_info_.dwProcessId;
@@ -327,7 +326,7 @@ ACE_Process_Options::getegid (void) const
 #endif /* ACE_WIN32 */
 
 ACE_INLINE ACE_TCHAR *
-ACE_Process_Options::command_line_buf (int *max_lenp)
+ACE_Process_Options::command_line_buf (size_t *max_lenp)
 {
   if (max_lenp != 0)
     *max_lenp = this->command_line_buf_len_;
